CIT-32 création de l'entité et dtos

+++ src/main/java/com/wora/citronix/dtos/champ/ChampCreateDTO.java
@@ -7,9 +7,10 @@ import java.util.UUID;
 @Data
 public class ChampCreateDTO {
     @Positive(message = "La superficie doit être positive.")
-    private double superficie;
+    private Long superficie;
 
     @NotNull(message = "L'identifiant de la ferme est obligatoire.")
     private UUID fermeId;
+
 }
